moved Searchbar into a dedicated folder and updated deprecated inputprops to slotprops
